+++ project/src/main/resources/static/css/cart.css
@@ -28,6 +28,7 @@
     bottom:0;               /*모든 방향에 0을 주면 화면에 꽉차게 됩니다.*/
     display:none;  /*이벤트가 발생할 때 띄우기 위해 숨김*/
     padding:15px; 
+    z-index: 100;
   }
 
   /*모달의 몸체*/
