+++ project/src/main/resources/static/js/itemList.js
@@ -14,49 +14,10 @@ $(function(){
         $("#item-popup").css('display','flex').hide().fadeIn();
         //팝업을 flex속성으로 바꿔준 후 hide()로 숨기고 다시 fadeIn()으로 효과
     });
-    
-	$("#cart-popup").click(function(){
-        $("#cart-modal").css('display','flex').hide().fadeIn();
-        //팝업을 flex속성으로 바꿔준 후 hide()로 숨기고 다시 fadeIn()으로 효과
-    });
-    
-	
-    
-    $("#close").click(function(){
+
+    $("#item-close").click(function(){
 		console.log('클릭은 되나')
         modalClose(); //모달 닫기 함수 호출
-    });
-    
-    
-    
-    $("#cart-popup .confirm").click(async function(){
-        modalClose(); //모달 닫기 함수 호출
-        console.log("여긴 cart 수정부분")
-        
-        const cartItems = await $.get("/cart/list");
-        const data = cartItems.map((item) => {
-			return {
-				amount : parseInt($(this).parent().parent().find(`tr#${item.sno}`).find(".amount").val()),
-				ino : item.ino,
-				sno : item.sno,
-				mno : item.mno
-			}
-		});
-		
-		console.log(data);
-		
-		$.ajax({
-			url: "/cart/items",
-			method: "PUT",
-			data: JSON.stringify(data),
-			contentType: "application/json",
-			success: function(result){
-				console.log(result);
-			},
-			error: function(err){
-				alert("잠시후에 다시 실행해주시길 바랍니다.")
-			}
-		})
     });
     
     $("#item-popup #confirm").click(function(){
@@ -75,6 +36,7 @@ $(function(){
 			}
 		})       
     });
+    
     $("#thumbnail").on("change", function(){
 		imageStr = '';
 		console.log("capture change event");
@@ -126,153 +88,7 @@ $(function(){
 
 
 function modalClose(){
-    $("#cart-popup").fadeOut(); //페이드아웃 효과
     $("#item-popup").fadeOut(); //페이드아웃 효과
-    $("#popup").fadeOut(); //페이드아웃 효과    
-}
-
-function cartModal(){
-	const $cart= $("#cart-popup table");
-	const $count = $("#cartCount");
-	const $checkPrice = $("#txt_getChkList");
-	$.ajax({
-		url: "/cart/items",
-		method: "GET",
-		dataType: "json",
-		success: function(result){
-			console.log(result);
-			if(result.length == 0){
-				$("#cart-popup .popup .popup-body").html(`
-					<div class="noneCart">
-						<h1>장바구니가 텅 비었습니다.</h1>
-					</div>
-				`)
-			}else{
-				
-				$cart.html( `
-					<tr>
-						<th class="selectall">
-	                        <input type="checkbox" name="cartitem" id="cbx_chkAll" value="selectall">
-	                    </th>
-	                    <th class="productinfoheader">상품 정보</th>
-	                    <th class="amountheader">수량</th>
-	                    <th class="amountallheader">주문금액</th>
-                    </tr>
-				` );
-				const cartHtml = result.map(([{thumbnail, name, brand, price}, cartDTO]) => {
-					let images = '';
-					if(thumbnail.split(',').length != 1){
-						images = thumbnail.split(',').map((URL) => 
-							`<img class="productimg" src="/display?fileName=${URL}&folderType=item">`
-						).join('');
-					} else {
-						if(thumbnail.split(':').length != 1){
-							images = `<img class="productimg" src="${thumbnail}">`;
-						}else{
-							images = `<img class="productimg" src="/display?fileName=${thumbnail}&folderType=item">`;
-						}
-					}
-					return `
-						<tr id="${cartDTO.sno}">
-                            <td class="select">
-                                <input type="checkbox" name="chk" value="select">
-                            </td>
-
-                            <td class="productinfobox">
-                                <div class="productimgbox">
-                           `
-                            + images + 
-                           `
-                                </div>
-                                <div class="producttextbox">
-                                    <div class="productname">${name}</div>
-                                    <div class="brandname">${brand}</div>
-                                    <div class="price">${price}원</div>
-                                </div>
-                            </td>
-
-                            <td class="amountbox">
-                                <input type="number" class="amount" value="${cartDTO.amount}"></div>
-                            </td>
-
-                            <td class="amountallbox">
-                                <div class="amountall">${cartDTO.amount * price}원</div>
-                            </td>
-                        </tr>
-					`
-				}).join("")
-				
-				$cart.append(cartHtml);
-				$count.html(`총 ${result.length}개의 상품`)
-			}
-			checkPrice($checkPrice);
-			$("#cart-modal-open #confirm").show();
-	        $("#cart-popup").css('display','flex').hide().fadeIn();
-		    $("#cbx_chkAll").click(function() {
-				console.log("여기 들어감")
-				if($("#cbx_chkAll").is(":checked")) $("input[name=chk]").prop("checked", true);
-				else $("input[name=chk]").prop("checked", false);
-				checkPrice($checkPrice);
-			});
-			$("input[name=chk]").click(function() {
-				var total = $("input[name=chk]").length;
-				var checked = $("input[name=chk]:checked").length;
-		
-				if(total != checked) $("#cbx_chkAll").prop("checked", false);
-				else $("#cbx_chkAll").prop("checked", true); 
-				checkPrice($checkPrice);
-			});
-			$("#cart-popup input[type=number]").bind("keyup mouseup", function(){
-				const price = parseInt($(this).parent().prev().find(".price").html());
-				const amount = $(this).val();
-				
-				$(this).parent().next().find(".amountall").html(`${price * amount}원`);
-				checkPrice($checkPrice);
-			})
-		},
-		error: function(err){
-			alert("로그인 후 이용하실 수 있습니다.");
-		}
-	})		
-};
-
-function deleteCart(){
-	let snos = new Array();
-	const items = $("#cart-popup").find("table tr").map(function(){
-		if($(this).find("input[name=chk]").is(":checked") == true){
-			snos.push(parseInt($(this).attr('id')));			
-		}
-	});
-	console.log(snos);
-	$.ajax({
-		url: "/cart/items",
-		method: "DELETE",
-		contentType: "application/json",
-		data: JSON.stringify(snos),
-		success: function(result){
-			cartModal();
-		}
-	})
-}
-
-function orderCart(){
-	let snos = new Array();
-	const items = $("#cart-popup").find("table tr").map(function(){
-		if($(this).find("input[name=chk]").is(":checked") == true){
-			snos.push(parseInt($(this).attr('id')));			
-		}
-	});
-	console.log(snos);
-	$.ajax({
-		url: "/history/items",
-		method: "POST",
-		contentType: "application/json",
-		data: JSON.stringify(snos),
-		success: function(result){
-			console.log("orderCart 등록 성공?")
-			deleteCart();
-		}
-	})
 }
 
 function checkPrice($el){
