+++ project/src/main/resources/templates/fragments/item_sec.html
@@ -73,6 +73,16 @@
 		<div class="sec3 box3" id="list"></div>
         <div class="pagingEl"></div>
 	</section>
+	<!-- 아이템 리스트 메인용 -->
+	<section th:fragment="mainFragment">
+		<script type="text/javascript" src="../js/main.js" defer></script>
+			<link rel="stylesheet" type="text/css" href="../css/cart.css">
+    <link rel="stylesheet" type="text/css" href="../css/camperproduct.css">
+    <link rel="stylesheet" type="text/css" href="../css/reviewwrite.css">
+	
+		<div class="sec3 box3" id="list"></div>
+        <div class="pagingEl"></div>
+	</section>
 	
 	<!-- 아이템 추가 모달 -->
 	<section th:fragment="list3Fragment">
