+++ project/src/main/resources/templates/index.html
@@ -2,47 +2,9 @@
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
 
 <head>
-	<meta charset="UTF-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>main</title>
-	<!--구글 폰트 링크-->
-	<link rel="preconnect" href="https://fonts.googleapis.com">
-	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-	<link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR&family=Varela+Round&display=swap"
-		rel="stylesheet">
-	<link rel="stylesheet"
-		href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-		
-	<!-- jquery -->
-	<script src="http://code.jquery.com/jquery-latest.js"></script>
-	
-	<!--typeit(페이지에 글씨입력되는것처럼 보이는 것)-->
-	<script src="https://unpkg.com/typeit@8.7.1/dist/index.umd.js" loop="true"></script>
-
-	<!--scrollOut(스크롤 내릴때 효과주는것)
-	<script src="https://unpkg.com/scroll-out/dist/scroll-out.min.js"></script> -->
-
-	<!--스크롤 내릴때 요소마다 다르게 주는 스크립트
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/rellax/1.12.1/rellax.min.js"></script> -->
-
-	<!--swiper-->
-	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@9/swiper-bundle.min.css" />
-	
-	<!-- Swiper JS -->
-	<script src="https://cdn.jsdelivr.net/npm/swiper@9/swiper-element-bundle.min.js"></script>
-	
-	<!--폰트 그림 링크-->
-	<script src="https://kit.fontawesome.com/bb87a5faf0.js" crossorigin="anonymous"></script>
-	
-	<!--css 연결-->
-	<link rel="stylesheet" type="text/css" href="../css/main.css">
-	<link rel="stylesheet" type="text/css" href="../css/header.css">
-	
-	<!--js연결-->
-	<script type="text/javascript" src="../js/main.js" defer></script>
-
-
+	<div th:replace="~{fragments/head_url :: headFragment}"></div>
+	<link rel="stylesheet" type="text/css" href="../css/cart.css">
+    <script type="text/javascript" src="../js/cartPopup.js" defer></script>	
 </head>
 
 <body>
@@ -78,7 +40,7 @@
 				<!--메인 바디 가운데 공간 안에 섹션 1, 2, 3이 나뉨-->
 				<div class="body_center box2">
 
-					<div th:replace="~{/fragments/item_sec :: list2Fragment}"></div>
+					<div th:replace="~{/fragments/item_sec :: mainFragment}"></div>
 					
 					<!--섹션나누기2-->
 					<div th:replace="~{/fragments/main_sec :: section2Fragment}"></div>
